pyproject.toml
[tool.poetry]
name = "django_json_null_expression_repro"
version = "0.1.0"
description = ""
authors = ["Daniel Schaffer <dschaffer@instawork.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2.10"
django-mysql = "^4.6.0"
pytest-django = "^4.5.2"
mysqlclient = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings"
